+++ engine/platform/unix/unix_main.cpp
@@ -532,7 +532,7 @@ char *Sys_ConsoleInput( void )
 				tty_FlushIn();
 				return NULL;
 			}
-			if ( tty_con.cursor >= sizeof( text ) - 1 )
+			if ( (size_t)tty_con.cursor >= sizeof( text ) - 1 )
 				return NULL;
 			// push regular character
 			tty_con.buffer[ tty_con.cursor ] = key;
@@ -669,7 +669,7 @@ static const struct Q3ToAnsiColorTable_s
 
 static const char *getANSIcolor( char Q3color ) {
 	int i;
-	for ( i = 0; i < ARRAY_LEN( tty_colorTable ); i++ ) {
+	for ( i = 0; (size_t)i < ARRAY_LEN( tty_colorTable ); i++ ) {
 		if ( Q3color == tty_colorTable[ i ].Q3color ) {
 			return tty_colorTable[ i ].ANSIcolor;
 		}
